# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Atelier Commerce API",
    "version": "1.0.0",
    "description": "Single-currency BDT commerce API. Bearer authentication; cash on delivery; server-side ownership and transactional inventory."
  },
  "servers": [
    {
      "url": "http://localhost:8081"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/products": {
      "get": {
        "summary": "Search active products",
        "operationId": "getApiProducts",
        "tags": [
          "Products"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "newest",
                "priceAsc",
                "priceDesc",
                "name"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "size",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 48
            }
          }
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Read active product",
        "operationId": "getApiProductsId",
        "tags": [
          "Products"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ]
      }
    },
    "/api/categories": {
      "get": {
        "summary": "List active product categories",
        "operationId": "getApiCategories",
        "tags": [
          "Categories"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/admin/products": {
      "get": {
        "summary": "Admin: list all products including hidden",
        "operationId": "getApiAdminProducts",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ]
      },
      "post": {
        "summary": "Admin: create product",
        "operationId": "postApiAdminProducts",
        "tags": [
          "Admin"
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          }
        }
      }
    },
    "/api/admin/products/{id}": {
      "put": {
        "summary": "Admin: replace product fields",
        "operationId": "putApiAdminProductsId",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ]
      },
      "delete": {
        "summary": "Admin: deactivate product",
        "operationId": "deleteApiAdminProductsId",
        "tags": [
          "Admin"
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "message"
        ]
      },
...
      "ProductInput": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "description": "Required when updating. Send the version returned when the product was loaded. A stale version returns 409. Ignored when creating."
          },
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "price": {
            "type": "number",
            "minimum": 0.01,
            "maximum": 9999999.99,
            "multipleOf": 0.01
          },
          "stock": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          },
          "category": {
            "type": "string",
            "maxLength": 60
          },
          "image": {
            "type": "string",
            "description": "HTTPS URL or /assets/ path"
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name",
          "description",
          "price",
          "stock",
          "category",
          "image",
          "active"
        ]
      },
      "ProductUpdate": {
        "allOf": [
          { "$ref": "#/components/schemas/ProductInput" },
          { "type": "object", "required": ["version"] }
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "price": {
            "type": "number",
            "minimum": 0.01,
            "maximum": 9999999.99,
            "multipleOf": 0.01
          },
          "stock": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          },
          "category": {
            "type": "string",
            "maxLength": 60
          },
          "image": {
            "type": "string",
            "description": "HTTPS URL or /assets/ path"
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "ProductPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "page": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "totalElements": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}